requirements.txt
pyTelegramBotAPI
python-dotenv
APScheduler